+++ classes/Hooks.php
@@ -11,15 +11,21 @@
 namespace HeimrichHannot\SearchPlus;
 
 
+use Contao\Config;
+
 class Hooks
 {
 	public function indexPageHook($strContent, $arrData, $arrSet)
 	{
-		if (preg_match_all('/href="(?<links>[^\"<]+\.pdf)"/i', $strContent, $matches))
-		{
-			Search::indexFiles($matches['links'], $arrSet);
-		}
-		else if(Validator::isRebuildIndexRequest())
+	    if (true === Config::get('search_enablePdfIndexing'))
+        {
+            if (preg_match_all('/href="(?<links>[^\"<]+\.pdf)"/i', $strContent, $matches))
+            {
+                Search::indexFiles($matches['links'], $arrSet);
+                return;
+            }
+        }
+		if(Validator::isRebuildIndexRequest())
 		{
 			Search::truncate($arrSet);
 		}