+++ languages/en/tl_settings.php
@@ -13,10 +13,10 @@ $arrLang = &$GLOBALS['TL_LANG']['tl_settings'];
 $arrLang['search_plus_legend'] = 'Search Plus Einstellungen';
 
 $arrLang['search_pdfMaxParsingSize'] = [
-    'Maximale PDF-Dateigröße für PDF-Parser',
-    'Die maximale Dateigröße der PDFs, welche vom Parser bearbeitet werden dürfen, um eine Speicherüberschreitung oder einen Timeout zu verhindern. Angabe in KiB. 0 für keine Größenbeschränkung. 1024KiB = 1MB'
+    'Maximum pdf size for pdf parser',
+    'The maximum file size of a pdf that can be processed by the pdf parser, to prevent memory overflow or process timeout. Specify in KiB. 0 means no file size limit. 1024KiB = 1MB'
 ];
 $arrLang['search_enablePdfIndexing'] = [
-    'PDF-Indexierung aktivieren',
-    'Hier kann die Indexierung von PDF für die Suche aktiviert oder deaktiviert werden.'
+    'Enable PDF-Indexing',
+    'Enable or disable the indexing of pdf files for search.'
 ];